refactor: 사용자 관련 모듈 및 엔티티 리팩토링, DbModule 추가 및 UsersModule 제거
