+++ src/application/DB/db.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { MongooseModule } from '@nestjs/mongoose';
+import { User, UserSchema } from './entity/user.entity';
+import { UserQuery } from './query/user.query';
+
+@Module({
+  imports: [
+    MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
+  ],
+  providers: [UserQuery],
+  exports: [UserQuery],
+})
+export class DbModule {}
