--- src/application/DB/users/users.module.ts
@@ -1,14 +1,0 @@
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from './schemas/user.schema';
-import { UsersService } from './users.service';
-
-@Module({
-  imports: [
-    MongooseModule.forFeature([{ name: User.name, schema: UserSchema }]),
-  ],
-  controllers: [],
-  providers: [UsersService],
-  exports: [UsersService],
-})
-export class UsersModule {}
